DT-6387 keep mqtt running while carousel shows other views, keep markerstatus saved between map renders
